added checkpointing structure to OcelotConfiguration
Former-commit-id: 4e0498bb53654d58d0f1b1e32dc6fb88a096be4d

+++ branches/ocelot-refactored/ocelot/cuda/implementation/CudaRuntime.cpp
@@ -131,7 +131,7 @@ static executive::dim3 convert(dim3 dim) {
 }
 
 static executive::dim3 convert(cudaPos pos) {
-	executive::dim3 ed3 = {pos.x, pos.y, pos.z};
+	executive::dim3 ed3 = {(int)pos.x, (int)pos.y, (int)pos.z};
 	return ed3;
 }
